Move db & tpm layers to separate directory
Common codes in the verifier/registrar_common will
be extracted after the layout is done.
