+++ keylime/db/keylime_db.py
@@ -1,6 +1,7 @@
-from keylime import keylime_logging
 from sqlalchemy.orm import sessionmaker, scoped_session
 from sqlalchemy.exc import SQLAlchemyError
+
+from keylime import keylime_logging
 
 logger = keylime_logging.init_logging('sql_session_manager')
