+++ keylime/db/registrar_db.py
