+++ keylime/tpm/tpm_abstract.py
