+++ test/test_registrar_db.py
@@ -1,6 +1,6 @@
 import unittest
-from keylime.registrar_db import RegistrarMain
-from keylime.keylime_database import SessionManager
+from keylime.db.registrar_db import RegistrarMain
+from keylime.db.keylime_db import SessionManager
 from sqlalchemy import create_engine
 
 # BEGIN TEST DATA
